Set Source IP of containers and optionally lift ICC restrictions
Add two environment variables "ICC_OVERRIDE" and "SNAT" which may be
set per container.

Setting "ICC_OVERRIDE" when running with global --icc=false allows the
container to talk to any other container. Allowing this container to be
talked to by other containers is subject to some future work.

Setting "SNAT" to some IP address installs an SNAT rule using --to-source,
so the container appears under this IP address when communicating
with the outside world.